Set final year for copyright by Facebook, Inc. on all files
Facebook has not developed this project since 2017, and aside
from the commit to adjust the licensing, there has been no work
on this project by them. As this is a new fork with no expectation
that Facebook will continue developing it, this adjustment is safe.

If they contribute to the project in the future, the year can be
updated again.